+++ contracts/books/bos/interfaces/IBOSCalculator.sol
@@ -6,7 +6,7 @@
 pragma solidity ^0.4.24;
 
 interface IBOSCalculator {
-    function membersOfClass(uint8 class) external view returns (uint32[]);
+    function membersOfClass(uint8 class) external view returns (uint40[]);
 
     function sharesOfClass(uint8 class) external view returns (bytes32[]);
